+++ BotManager.py
@@ -11,7 +11,7 @@ import subprocess
 import json
 from telepot.delegate import per_chat_id, create_open
 
-from telepot.namedtuple import ReplyKeyboardMarkup, KeyboardButton, ReplyKeyboardHide, ForceReply 
+from telepot.namedtuple import ReplyKeyboardMarkup, KeyboardButton, ReplyKeyboardRemove, ForceReply 
 from telepot.namedtuple import InlineKeyboardMarkup, InlineKeyboardButton 
 from telepot.namedtuple import InlineQueryResultArticle, InlineQueryResultPhoto, InputTextMessageContent