+++ CHANGELOG.md
@@ -1,5 +1,19 @@
 ### 변경 내역
 
+**0.5.6 (2017-01-02)**
+
+DSM 6.x 전용 bot.sh 파일 추가
+
+DSM 6.x 에서는 최초 설치 후 bot6.sh.ori 파일을 bot.sh 로 복사 하셔서 사용하시면 됩니다.
+
+telepot 이 업데이트 되면서 ReplyKeyboardHide 이름이 ReplyKeyboardRemove 이름으로 변경 되어 소스상에 업데이트 하였습니다.
+
+pip install telepot --upgrade 를 수행 하신 다음 xpebot 을 업데이트 하셔서 사용하시면 됩니다.
+
+토렌트 파일 전송시 파일 이름에 특정 내용이 있을 경우 문자열 인코딩 실패 오류를 수정하였습니다.
+
+------
+
 **0.5.5 (2016-10-26)**
 
 DownloadStation 에서 Torrent 제목에 대괄호([])가 포함 되어 있고 대괄호 사이 공백이 있는 경우 메시지 전송을 못하는 문제 수정